Start using a version range for lsp-types
This switches us to use a version range (`>=0.70,<0.74`) for our dependency on `lsp-types`, seeing as we only use a couple of types from this repository (`lsp_types::{Position, Range}`) that don’t change very often. This will mean that we can make releases of `codespan-lsp` that update the maximum version _without_ breaking backwards compatibility for consumers.